requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mailman>=0.3.0
Flask-Migrate==4.0.5
Flask-Limiter==3.11.0
Flask-WTF==1.2.1
PyJWT==2.8.0
PyMySQL==1.1.0
cryptography==41.0.7
httpx[http2]>=0.25.0
SQLAlchemy==2.0.35
email-validator>=2.1.1
Pillow>=10.0.0
pillow-heif>=0.18.0
python-dotenv==1.0.0
hypothesis==6.92.1
pytest==7.4.3
pytest-flask==1.3.0
factory-boy>=3.3.0
markdown>=3.5.0
